Move DiscIO enums to a new file
At first there weren't many enums in Volume.h, but the number has been
growing, and I'm planning to add one more for regions. To not make
Volume.h too large, and to avoid needing to include Volume.h in code
that doesn't use volume objects, I'm moving the enums to a new file.
I'm also turning them into enum classes while I'm at it.

+++ Source/Core/DolphinQt2/GameList/GameListModel.cpp
@@ -38,7 +38,7 @@ QVariant GameListModel::data(const QModelIndex& index, int role) const
     case COL_SIZE:
       return game->GetFileSize();
     case COL_COUNTRY:
-      return game->GetCountryID();
+      return static_cast<int>(game->GetCountryID());
     case COL_RATING:
       return game->GetRating();
     }
